I think I figured out why the RAM usage was increasing - it was actually the python script lol. Still need to figure out random crashes, and fix train movement a lil more

+++ t_renderer.h
@@ -25,6 +25,6 @@ typedef struct pixelDataExtended
 
 bool t_rend_init(SDL_Window **window, SDL_Renderer **renderer, imageData *imgData, pixelDataExtended ***tBuffer);
 bool t_rend_event();
-bool t_rend_drawPixels(imageData *imgData, SDL_Renderer *renderer, trainNode *tNodes, int16_t tNodeSize, pixelDataExtended **tBuff);
+bool t_rend_drawPixels(imageData *imgData, SDL_Renderer *renderer, trainNode *tNodes, uint8_t tNodeSize, pixelDataExtended **tBuff);
 void t_rend_quit(SDL_Window *window, SDL_Renderer *renderer, imageData *imgData, pixelDataExtended ***tBuffer);
